fixed require cycle warnings (refactoring imports)
